mac/app/requirements.txt
# Google 表格与 Drive
gspread>=6.0.0
gspread-formatting>=1.1.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# 下载 Instagram 视频
yt-dlp>=2024.5.0  # 2024.5+ 支持小红书

# 环境变量
python-dotenv>=1.0.0

# 飞书 API（可选，用于把封面文件写入飞书多维表格「文件」列）
requests>=2.28.0
